refactor: TripSearchPanel -> TripSearchListPanel 리네이밍
