+++ src/components/trip/TripSearchListPanel.vue
